Split composer tests into multiple files
Change-type: patch
